requirements.txt
openai
PyMuPDF
tqdm
python-dotenv
